% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addsubplotleg.r
\name{addsubplotleg}
\alias{addsubplotleg}
\title{addsubplotleg}
\usage{
addsubplotleg(lab, cex = 1.5, ...)
}
\arguments{
\item{lab}{a string you want to display}

\item{cex}{the text cex}
}
\description{
adds a pretty subplot legend e.g., (a), (b), ..., etc. for par multipanel plots
}
\examples{
\dontrun{
par(mfrow = c(2, 1)
plot(1:10)
addsubplotleg('(a)')
plot(21:30)
addsubplotleg('(b)')
}
}
