% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_alpha.R
\name{change_alpha}
\alias{change_alpha}
\title{change (or add) alpha to colors}
\usage{
change_alpha(col, alpha)
}
\arguments{
\item{col}{color or vector of colors}

\item{alpha}{value to add or change. on [0, 1].}
}
\value{
rgb colors or vector of rgbcolors with alpha
}
\description{
change (or add) alpha to colors
}
\examples{
# numeric
change_alpha(1, .5)

# rgb
change_alpha(rgb(1, 0, 1), .5)
change_alpha("#ff00ff", .5)

# multiple
change_alpha(1:3, alpha = 0.5)
change_alpha(1:3, alpha = c(0.1, 0.2, 0.3))
}
