% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chau.R
\name{chau}
\alias{chau}
\title{apply Chauvenet's criterion to find outliers}
\usage{
chau(y, threshold, na.rm = TRUE)
}
\arguments{
\item{y}{a numeric vector time series}

\item{na.rm}{boolean defaults to TRUE for \code{mean} and \code{sd}}

\item{predetermined}{threshold}
}
\value{
a list of two vectors: \code{outliers} boolean vector where TRUE is the outliers and \code{crit} which are the critical values.
}
\description{
apply Chauvenet's criterion to find outliers
}
\examples{
# make some data
y <- sin(1:100 / pi)
y[25] <- y[25]*3

chau(y, .5)
}
