\name{cioffihelperfunctions-package}
\alias{cioffihelperfunctions-package}
\alias{cioffihelperfunctions}
\docType{package}
\title{
\packageTitle{cioffihelperfunctions}
}
\description{
\packageDescription{cioffihelperfunctions}
}
\details{
date2num
num2date
dateseq
matchtimes

getstretchid

latlond.hav
torad

trubetskoy_colors

rcsv
wcsv
}
\author{
if you're actually here you should know how to find me.
}

