% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date2num.R
\name{date2num}
\alias{date2num}
\title{translate character dates to numeric dates}
\usage{
date2num(d, tz = "UTC", ...)
}
\arguments{
\item{d}{a character vector of dates}

\item{tz, format}{as expected by \code{\link[base]{as.POSIXct}}.}

\item{...}{any other good stuff you want to pass to \code{as.POSIXct}.}
}
\description{
convert character dates to numeric dates (seconds since UNIX epoch by default).
}
\note{
beware if you open any csvs in excel and re-save, the dates will likely be put into an absurb ambiguous form. additionally, seconds tend to be obliterated.
}
\examples{
date2num("1988-03-12 07:00:00", tz = "UTC", format = "\%Y-\%m-\%d \%H:\%M:\%S")
}
\seealso{
\code{\link[base]{as.POSIXct}}

Other date manipulators: \code{\link{num2date}}
}
\concept{date manipulators}
