% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2mat.r
\name{df2mat}
\alias{df2mat}
\title{df2mat}
\usage{
df2mat(d, symmetrical = TRUE)
}
\arguments{
\item{d}{dataframe to convert has to have 3 columns with the data in the third}

\item{symmetrical}{bool only symmetrical supported right now}
}
\value{
matrix
}
\description{
convert a matrix to a dataframe
}
\seealso{
\code{\link[chf]{mat2df}}
}
