% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formenu.r
\name{formenu}
\alias{formenu}
\title{tiny function for navigating pages of plots mainly}
\usage{
formenu(f = function(i) {
    print(i)
})
}
\arguments{
\item{f}{this is the function you want to iterate}
}
\description{
tiny function for navigating pages of plots mainly
}
\details{
Sometimes I have a big for loop with a bunch of diagnostic plots and a readline() at the end. Use this instead you can go forward and back and skip all around. Some basic error checking.
}
