% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makestretchid.R
\name{getstretchid}
\alias{getstretchid}
\title{make a stretch id from a date or something}
\usage{
getstretchid(stretches)
}
\arguments{
\item{stretches}{a numeric or character vector to convert into a stretchid}
}
\value{
a stretchid numeric vector starting at 1
}
\description{
make a stretch id from a date or something
}
\details{
i do this all the time, but probably there is a cleverer pattern where you don't need to do it that i don't know about
}
\examples{
# consider a date vector
v1 <- c("2019-01-01", "2019-01-01", "2019-01-02", "2019-01-03")

# get the stretches
getstretchid(v1)
}
