% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invlogit.R
\name{invlogit}
\alias{invlogit}
\title{inverse logit}
\usage{
invlogit(a)
}
\arguments{
\item{a}{the numberic vector to which the deed is done}
}
\value{
numeric vector deed done
}
\description{
does what you want
}
\examples{
anum <- 0.48
logit_anum <- log(anum/(1 - anum))

invlogit(logit_anum)
}
\seealso{
\code{\link[cioffihelperfunctions]{logit}}
}
