% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelgenerator.R
\name{labelgenerator}
\alias{labelgenerator}
\title{automatically generate some letter labels for arbitrary n}
\usage{
labelgenerator(n, case = c("upper", "lower"))
}
\arguments{
\item{n}{how many labels you want}

\item{case}{upper or lower case letters}
}
\value{
a vector of labels
}
\description{
automatically generate some letter labels for arbitrary n
}
\details{
basically the column labels in a spreadsheet
}
\examples{
# make some labels
labs <- labelgenerator(10)
plot(1:10)
axis(3, at = 1:10, lab = labs)

# they'll repeat in the way you expect
labelgenerator(1000)
}
