% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat2df.r
\name{mat2df}
\alias{mat2df}
\title{mat2df}
\usage{
mat2df(
  m,
  datlabel = "value",
  symmetrical = TRUE,
  diag = FALSE,
  stringsAsFactors = FALSE
)
}
\arguments{
\item{m}{matrix}

\item{datlabel}{character of what you want the column to be in the df}

\item{symmetrical}{bool only TRUE is currently supported}

\item{diag}{bool only FALSE is currently supported}

\item{stringsAsFactors}{bool don't mess with it}
}
\value{
data.frame
}
\description{
convert a matrix to a dataframe
}
\seealso{
\code{\link[chf]{df2mat}}
}
