% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchtimes.R
\name{matchtimes}
\alias{matchtimes}
\title{closest match for instance for times}
\usage{
matchtimes(t1, t2)
}
\arguments{
\item{t1}{a numeric vector to match}

\item{t2}{the second numeric vectgor to match}
}
\value{
another numeric vector of length equal to \code{t1}
}
\description{
closest match for instance for times
}
\details{
it rounds up, who cares?
}
\examples{
# consider two vectors
v1 <- seq(1, 11, by = 1)
v2 <- seq(1, 11, by = 4)

# matchitup
matchtimes(v1, v2)
}
