% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orcid_my.r
\name{orcid_my}
\alias{orcid_my}
\title{my orcid id e.g. for setting seeds}
\usage{
orcid_my(char = FALSE)
}
\arguments{
\item{char}{default FALSE returns a numeric (note this will remove any leading 0s) otherwise will return a char.}
}
\value{
numeric representation of my orcid
}
\description{
my orcid id e.g. for setting seeds
}
\examples{
\dontrun{
set.seed(orcid_wrc())
x1 <- rnorm(100)
}
}
