% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orcid_wrc.r
\name{orcid_wrc}
\alias{orcid_wrc}
\title{my orcid id e.g. for setting seeds}
\usage{
orcid_wrc(char = FALSE)
}
\value{
numeric representation of my orcid
}
\description{
my orcid id e.g. for setting seeds
}
\examples{
\dontrun{
set.seed(orcid_wrc())
x1 <- rnorm(100)
}
}
