% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trubetskoy_colors.R
\name{trubetskoy_colors}
\alias{trubetskoy_colors}
\title{use this color pallete}
\usage{
trubetskoy_colors(n, alpha = 1, nogray = FALSE)
}
\arguments{
\item{n}{the number of colors you want (1:22 if you want black and white and gray; 1:19 if not)}

\item{alpha}{set the alpha of all the colors together at once I don't let you do it indepdently}

\item{nogray}{defaults to \code{FALSE}, if set to \code{TRUE} then it doesn't return black and white amd gray.}
}
\value{
a vector of hexcoded rgb colors
}
\description{
use this color pallete
}
\details{
this uses the color pallete described here: \url{https://sashat.me/2017/01/11/list-of-20-simple-distinct-colors/} and inspired by Kenneth Kelly's 22 colors of maximum contrast. Kelly's are better matched for "brightness" but these are more distinct to my eyes.
}
\examples{
\dontrun{
plot(1:22, rep(1, 22), pch = 16, cex = 5, col = trubetskoy_colors(22))  
}
}
