% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrcorcid.r
\name{wrcorcid}
\alias{wrcorcid}
\title{my orcid id e.g. for setting seeds}
\usage{
wrcorcid()
}
\value{
numeric representation of my orcid
}
\description{
my orcid id e.g. for setting seeds
}
\examples{
\dontrun{
set.seed(wrcorcid())
x1 <- rnorm(100)
}
}
