\name{cioffihelperfunctions-package}
\alias{cioffihelperfunctions-package}
\alias{cioffihelperfunctions}
\docType{package}
\title{
\packageTitle{cioffihelperfunctions}
}
\description{
\packageDescription{cioffihelperfunctions}
}
\details{
matchtimes
}
\author{
if you're actually here you should know how to find me.
}